+++ Source/BugsnagEvent.m
@@ -24,6 +24,7 @@
 #import "Private.h"
 #import "BSG_RFC3339DateTool.h"
 #import "Private.h"
+#import "BugsnagKeys.h"
 
 @interface BugsnagBreadcrumb ()
 + (instancetype _Nullable)breadcrumbWithBlock:
@@ -271,7 +272,7 @@ static NSString *const DEFAULT_EXCEPTION_TYPE = @"cocoa";
 /**
  *  The release stages used to notify at the time this report is captured
  */
-@property(readwrite, copy, nullable) NSArray *notifyReleaseStages;
+@property(readwrite, copy, nullable) NSArray *enabledReleaseStages;
 
 /**
  *  Property overrides
@@ -324,7 +325,7 @@ static NSString *const DEFAULT_EXCEPTION_TYPE = @"cocoa";
                 }
             }
         } else {
-            _notifyReleaseStages = BSGLoadConfigValue(report, @"notifyReleaseStages");
+            _enabledReleaseStages = BSGLoadConfigValue(report, BSGKeyEnabledReleaseStages);
             _releaseStage = BSGParseReleaseStage(report);
             _threads = [report valueForKeyPath:@"crash.threads"];
             RegisterErrorData *data = [RegisterErrorData errorDataFromThreads:_threads];
@@ -397,7 +398,7 @@ initWithErrorName:(NSString *_Nonnull)name
         _overrides = [NSDictionary new];
         _metadata = metadata ?: [NSDictionary new];
         _releaseStage = config.releaseStage;
-        _notifyReleaseStages = config.notifyReleaseStages;
+        _enabledReleaseStages = config.enabledReleaseStages;
         // Set context based on current values.  May be nil.
         _context = metadata[BSGKeyContext] ?: [[Bugsnag configuration] context];
         NSMutableArray *crumbs = [NSMutableArray new];
@@ -529,8 +530,8 @@ initWithErrorName:(NSString *_Nonnull)name
 }
 
 - (BOOL)shouldBeSent {
-    return [self.notifyReleaseStages containsObject:self.releaseStage] ||
-           (self.notifyReleaseStages.count == 0 &&
+    return [self.enabledReleaseStages containsObject:self.releaseStage] ||
+           (self.enabledReleaseStages.count == 0 &&
             [[Bugsnag configuration] shouldSendReports]);
 }
 
@@ -781,6 +782,10 @@ initWithErrorName:(NSString *_Nonnull)name
     }
 
     BSGArrayAddSafeObject(bugsnagThreads, threadDict);
+}
+
+- (BOOL)unhandled {
+    return self.handledState.unhandled;
 }
 
 @end